+++ src/gospatial/app/db.go
@@ -121,7 +121,7 @@ func (self *Database) startLogger() {
 	if err != nil {
 		Error.Fatal(err)
 	}
-	log_file := strings.Replace(dir, "bin", "db.log", -1)
+	log_file := strings.Replace(dir, "bin", "log/db.log", -1)
 	db_log, err := os.OpenFile(log_file, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0666)
 	if err != nil {
 		Error.Println("Error opening file: %v", err)
@@ -130,7 +130,7 @@ func (self *Database) startLogger() {
 			Error.Fatal("Error opening file: %v", err)
 		}
 	}
-	self.Logger = log.New(db_log, "[GOSPATIAL] DB | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
+	self.Logger = log.New(db_log, "[DB] | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 }
 
 /*=======================================*/
@@ -300,7 +300,7 @@ func (self *Database) InsertLayer(datasource string, geojs *geojson.FeatureColle
 		Error.Println(err)
 	}
 	//
-	self.Logger.Println(`{"method": "insert_layer", "data":` + string(value) + `}`)
+	self.Logger.Println(`{"method": "insert_layer", "datasource":` + datasource + `, "data":` + string(value) + `}`)
 	// Insert layer into database
 	Debug.Printf("Database insert datasource [%s]", datasource)
 	err = conn.Update(func(tx *bolt.Tx) error {
@@ -340,7 +340,7 @@ func (self *Database) InsertLayers(datsources map[string]*geojson.FeatureCollect
 		if err != nil {
 			Error.Println(err)
 		}
-		self.Logger.Println(`{"method": "insert_layer", "data":` + string(value) + `}`)
+		self.Logger.Println(`{"method": "insert_layer", "datasource":` + datasource + `, "data":` + string(value) + `}`)
 		// Insert layer into database
 		Debug.Printf("Database insert datasource [%s]", datasource)
 		err = conn.Update(func(tx *bolt.Tx) error {