+++ src/gospatial/app/generalHandlers.go
@@ -32,14 +32,14 @@ func MapHandler(w http.ResponseWriter, r *http.Request) {
 	if SuperuserKey == apikey {
 		map_tmpl := "./templates/map_admin.html"
 		tmpl, _ := template.ParseFiles(map_tmpl)
-		Info.Println(r.RemoteAddr, "GET /map/"+ds+" [200]")
+		network_logger_Info.Println(r.RemoteAddr, "GET /map/"+ds+" [200]")
 		tmpl.Execute(w, MapData{Datasource: ds})
 	}
 
 	/*=======================================*/
 	// Check for apikey in request
 	if apikey == "" {
-		Error.Println(r.RemoteAddr, "POST /api/v1/layer/"+ds+"/feature [401]")
+		network_logger_Error.Println(r.RemoteAddr, "POST /api/v1/layer/"+ds+"/feature [401]")
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
 		return
 	}
@@ -47,14 +47,14 @@ func MapHandler(w http.ResponseWriter, r *http.Request) {
 	// Get customer from database
 	customer, err := DB.GetCustomer(apikey)
 	if err != nil {
-		Warning.Println(r.RemoteAddr, "POST /api/v1/layer/"+ds+"/feature [404]")
+		network_logger_Warning.Println(r.RemoteAddr, "POST /api/v1/layer/"+ds+"/feature [404]")
 		http.Error(w, err.Error(), http.StatusNotFound)
 		return
 	}
 
 	// Check customer datasource list
 	if !stringInSlice(ds, customer.Datasources) {
-		Error.Println(r.RemoteAddr, "POST /api/v1/layer/"+ds+"/feature [401]")
+		network_logger_Error.Println(r.RemoteAddr, "POST /api/v1/layer/"+ds+"/feature [401]")
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
 		return
 	}
@@ -63,7 +63,7 @@ func MapHandler(w http.ResponseWriter, r *http.Request) {
 	// Return results
 	map_tmpl := "./templates/map_admin.html"
 	tmpl, _ := template.ParseFiles(map_tmpl)
-	Info.Println(r.RemoteAddr, "GET /map/"+ds+" [200]")
+	network_logger_Info.Println(r.RemoteAddr, "GET /map/"+ds+" [200]")
 	tmpl.Execute(w, MapData{Datasource: ds, Apikey: apikey})
 
 }