+++ src/gospatial/app/socketHandlers.go
@@ -105,7 +105,7 @@ func serveWs(w http.ResponseWriter, r *http.Request) {
 	ip := r.RemoteAddr
 	ws, err := upgrader.Upgrade(w, r, nil)
 	if err != nil {
-		Error.Println(r.RemoteAddr, "WS /ws/"+ds+" [500]")
+		network_logger_Error.Println(r.RemoteAddr, "WS /ws/"+ds+" [500]")
 		Error.Println(err)
 		return
 	}
@@ -118,7 +118,7 @@ func serveWs(w http.ResponseWriter, r *http.Request) {
 		Hub.Sockets[ds][conn.c] = ws
 		// Info.Println(r.RemoteAddr, "WS /ws/"+conn.ds+" [200]")
 	}
-	Info.Println(r.RemoteAddr, "WS /ws/"+conn.ds+" [200]")
+	network_logger_Info.Println(r.RemoteAddr, "WS /ws/"+conn.ds+" [200]")
 	Hub.broadcastAllDsViewers(false, conn.ds)
 	go messageListener(&conn)
 }