+++ test_runner.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-echo "Setting workspace"
-export GOPATH="`pwd`"
-
-cd src/gospatial/app
-go test -bench=. -test.benchmem
-